requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.14
Flask==1.1.1
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.16.0
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
tensorboard==1.15.0
tensorflow==1.14.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
urllib3==1.25.7
